Add platform optimization and workflow automation tools
- Introduced new tools for platform optimization including specifications retrieval, platform recommendations, video analysis, and optimization for specific platforms.
- Implemented batch optimization for multiple platforms and thumbnail generation tools.
- Added workflow automation tools for managing video editing processes, including workflow execution, status retrieval, cancellation, and custom workflow creation.
- Enhanced video analysis capabilities to suggest suitable workflows based on content characteristics.

+++ src-tauri/src/media/media_analyzer.rs
@@ -82,8 +82,8 @@ impl MediaAnalyzer {
       .find(|s| s.codec_type == "video")
     {
       resolution = Some((
-        video_stream.width.unwrap_or(0) as u32,
-        video_stream.height.unwrap_or(0) as u32,
+        video_stream.width.unwrap_or(0),
+        video_stream.height.unwrap_or(0),
       ));
 
       if let Some(fps_str) = &video_stream.r_frame_rate {
@@ -167,7 +167,7 @@ impl MediaAnalyzer {
         .iter()
         .find(|s| s.codec_type == "audio");
 
-      let channels = audio_stream.and_then(|s| s.channels).unwrap_or(2) as u8;
+      let channels = audio_stream.and_then(|s| s.channels).unwrap_or(2);
 
       let sample_rate = audio_stream
         .and_then(|s| s.sample_rate.as_ref())
